Remove NullGuardedConstraint.Assert
