+++ tests/FakeItEasy.Tests/Creation/CastleDynamicProxy/CastleDynamicProxyGeneratorTests.cs
@@ -4,9 +4,11 @@ namespace FakeItEasy.Tests.Creation.CastleDynamicProxy
     using System.Collections.Generic;
     using System.Diagnostics.CodeAnalysis;
     using System.Linq;
+    using System.Linq.Expressions;
     using FakeItEasy.Core;
     using FakeItEasy.Creation;
     using FakeItEasy.Creation.CastleDynamicProxy;
+    using FakeItEasy.Tests;
     using FluentAssertions;
     using NUnit.Framework;
     using TestHelpers;
@@ -234,8 +236,8 @@ namespace FakeItEasy.Tests.Creation.CastleDynamicProxy
             // Act
 
             // Assert
-            NullGuardedConstraint.Assert(() =>
-                this.generator.GenerateProxy(typeof(IInterfaceType), Enumerable.Empty<Type>(), null, A.Dummy<IFakeCallProcessorProvider>()));
+            Expression<Action> call = () => this.generator.GenerateProxy(typeof(IInterfaceType), Enumerable.Empty<Type>(), null, A.Dummy<IFakeCallProcessorProvider>());
+            call.Should().BeNullGuarded();
         }
 
         [Test]
