+++ tests/FakeItEasy.Tests/Expressions/ExpressionCallRuleTests.cs
@@ -1,9 +1,11 @@
 namespace FakeItEasy.Tests.Expressions
 {
     using System;
+    using System.Linq.Expressions;
     using FakeItEasy.Configuration;
     using FakeItEasy.Core;
     using FakeItEasy.Expressions;
+    using FakeItEasy.Tests;
     using FakeItEasy.Tests.TestHelpers;
     using FluentAssertions;
     using NUnit.Framework;
@@ -49,8 +51,8 @@ namespace FakeItEasy.Tests.Expressions
         [Test]
         public void Constructor_should_be_null_guarded()
         {
-            NullGuardedConstraint.Assert(() =>
-                new ExpressionCallRule(this.callMatcher));
+            Expression<Action> call = () => new ExpressionCallRule(this.callMatcher);
+            call.Should().BeNullGuarded();
         }
 
         [Test]
